# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FL-ReportCard",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/admin/import/dry-run": {
      "post": {
        "tags": [
          "admin-import"
        ],
        "summary": "Dry Run Import",
        "operationId": "dry_run_import",
        "parameters": [
          {
            "name": "year_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 160,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "grade_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 8,
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "Grade Level"
            }
          },
          {
            "name": "section",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 32
                },
                {
                  "type": "null"
                }
              ],
              "title": "Section"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "german",
                    "french",
                    "none"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Language"
            }
          },
          {
            "name": "action",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "new_student",
                    "rename",
                    "move",
                    "class_changed",
                    "language_change",
                    "unchanged",
                    "manual_added",
                    "language_edited"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_dry_run_import"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/import/commit": {
      "post": {
        "tags": [
          "admin-import"
        ],
        "summary": "Commit Import",
        "operationId": "commit_import",
        "parameters": [
          {
            "name": "year_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Year Id"
            }
          },
          {
            "name": "expected_sha256",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Expected Sha256"
            }
          },
          {
            "name": "expected_review_sha256",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "pattern": "^[0-9a-f]{64}$"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Review Sha256"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_commit_import"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreview"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_commit_import": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "class_moves": {
            "type": "string",
            "maxLength": 1000000,
            "title": "Class Moves",
            "default": "[]"
          },
          "roster_edits": {
            "type": "string",
            "maxLength": 2000000,
            "title": "Roster Edits",
            "default": "{}"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_commit_import"
      },
      "Body_dry_run_import": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "class_moves": {
            "type": "string",
            "maxLength": 1000000,
            "title": "Class Moves",
            "default": "[]"
          },
          "roster_edits": {
            "type": "string",
            "maxLength": 2000000,
            "title": "Roster Edits",
            "default": "{}"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_dry_run_import"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportClassMove": {
        "properties": {
          "school_number": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "School Number"
          },
          "grade_level": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 0.0,
            "title": "Grade Level"
          },
          "section": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Section"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "school_number",
          "grade_level",
          "section"
        ],
        "title": "ImportClassMove"
      },
      "ImportIssue": {
        "properties": {
          "level": {
            "type": "string",
            "title": "Level"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "sheet": {
            "type": "string",
            "title": "Sheet"
          },
          "cell": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cell"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "level",
          "code",
          "sheet",
          "message"
        ],
        "title": "ImportIssue"
      },
      "ImportLanguageChange": {
        "properties": {
          "school_number": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "School Number"
          },
          "language": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "german",
                  "french"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "school_number",
          "language"
        ],
        "title": "ImportLanguageChange"
      },
      "ImportPreview": {
        "properties": {
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "review_sha256": {
            "type": "string",
            "title": "Review Sha256"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/ImportIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/PreviewRow"
            },
            "type": "array",
            "title": "Rows"
          },
          "total_rows": {
            "type": "integer",
            "title": "Total Rows"
          },
          "filtered_rows": {
            "type": "integer",
            "title": "Filtered Rows"
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset"
          },
          "classes": {
            "items": {
              "$ref": "#/components/schemas/PreviewClass"
            },
            "type": "array",
            "title": "Classes"
          },
          "class_moves": {
            "items": {
              "$ref": "#/components/schemas/ImportClassMove"
            },
            "type": "array",
            "title": "Class Moves"
          },
          "roster_edits": {
            "$ref": "#/components/schemas/ImportRosterEdits"
          }
        },
        "type": "object",
        "required": [
          "sha256",
          "review_sha256",
          "counts",
          "issues",
          "rows",
          "total_rows",
          "filtered_rows",
          "next_offset",
          "classes",
          "class_moves"
        ],
        "title": "ImportPreview"
      },
      "ImportRosterEdits": {
        "properties": {
          "additions": {
            "items": {
              "$ref": "#/components/schemas/ImportStudentAdd"
            },
            "type": "array",
            "maxItems": 5000,
            "title": "Additions"
          },
          "removed_school_numbers": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Removed School Numbers"
          },
          "language_changes": {
            "items": {
              "$ref": "#/components/schemas/ImportLanguageChange"
            },
            "type": "array",
            "maxItems": 10000,
            "title": "Language Changes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ImportRosterEdits"
      },
      "ImportStudentAdd": {
        "properties": {
          "school_number": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "School Number"
          },
          "grade_level": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 0.0,
            "title": "Grade Level"
          },
          "section": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Section"
          },
          "full_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Full Name"
          },
          "language": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "german",
                  "french"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "school_number",
          "grade_level",
          "section",
          "full_name"
        ],
        "title": "ImportStudentAdd"
      },
...
      "PreviewClass": {
        "properties": {
          "grade_level": {
            "type": "integer",
            "title": "Grade Level"
          },
          "section": {
            "type": "string",
            "title": "Section"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "grade_level",
          "section",
          "count"
        ],
        "title": "PreviewClass"
      },
      "PreviewRow": {
        "properties": {
          "row": {
            "$ref": "#/components/schemas/RowModel"
          },
          "actions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Actions"
          },
          "original_grade_level": {
            "type": "integer",
            "title": "Original Grade Level"
          },
          "original_section": {
            "type": "string",
            "title": "Original Section"
          },
          "original_class": {
            "type": "string",
            "title": "Original Class"
          }
        },
        "type": "object",
        "required": [
          "row",
          "actions",
          "original_grade_level",
          "original_section",
          "original_class"
        ],
        "title": "PreviewRow"
      },
...
      "RowModel": {
        "properties": {
          "school_number": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "School Number"
          },
          "full_name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 2,
            "title": "Full Name"
          },
          "grade_level": {
            "type": "integer",
            "maximum": 8.0,
            "minimum": 0.0,
            "title": "Grade Level"
          },
          "section": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Section"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "language_present": {
            "type": "boolean",
            "title": "Language Present",
            "default": false
          },
          "sheet": {
            "type": "string",
            "title": "Sheet"
          },
          "row_number": {
            "type": "integer",
            "title": "Row Number"
          }
        },
        "type": "object",
        "required": [
          "school_number",
          "full_name",
          "grade_level",
          "section",
          "sheet",
          "row_number"
        ],
        "title": "RowModel"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}